# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Uni Ride Sharing API",
    "description": "API documentation for the university ride sharing application",
    "version": "1.0.0"
  },
  "host": "localhost:3002",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "sessionAuth": {
      "type": "apiKey",
      "name": "Cookie",
      "in": "header",
      "description": "Session cookie for authentication"
    }
  },
  "paths": {
...
    "/bookings": {
      "post": {
        "summary": "Create a booking",
        "description": "Books a seat on a ride",
        "tags": ["Bookings"],
        "security": [{"sessionAuth": []}],
        "parameters": [
          {
            "name": "bookingData",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["rideId"],
              "properties": {
                "rideId": {
                  "type": "integer",
                  "example": 123
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Booking created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Ride not found"
          },
          "409": {
            "description": "No seats available"
          }
        }
      }
    },
...
    "/bookings/{bookingId}/cancel": {
      "post": {
        "summary": "Cancel a booking",
        "description": "Cancels a booking",
        "tags": ["Bookings"],
        "security": [{"sessionAuth": []}],
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the booking"
          }
        ],
        "responses": {
          "200": {
            "description": "Booking cancelled"
          },
          "403": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Booking not found"
          }
        }
      }
    },
...
  }
} 